requirements.txt
streamlit>=1.23.1
python-dotenv>=0.19.0
Pillow>=9.0.0
google-generativeai>=0.8.0
requests>=2.27.0
asyncio>=3.4.3 